requirements.txt
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1640817743617/work
argon2-cffi-bindings @ file:///Users/runner/miniforge3/conda-bld/argon2-cffi-bindings_1649500376817/work
asgiref==3.5.0
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1618968359944/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
backports.zoneinfo==0.2.1
beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1649463573192/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1649361991009/work
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1636046166270/work
cryptography==36.0.2
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1649586398338/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
Django==4.0.4
django-sslserver==0.22
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1646044401614/work
fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1641751198313/work/dist
flit_core @ file:///home/conda/feedstock_root/build_artifacts/flit-core_1645629044586/work/source/flit_core
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1648728616176/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1649682376682/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1649684576414/work/dist/ipykernel-6.13.0-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1648413609726/work
ipython-genutils==0.2.0
ipywidgets @ file:///home/conda/feedstock_root/build_artifacts/ipywidgets_1647456365981/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1649067325908/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1648299710939/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1642000296051/work
jupyter @ file:///Users/runner/miniforge3/conda-bld/jupyter_1637233406932/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1649340884673/work
jupyter-console @ file:///home/conda/feedstock_root/build_artifacts/jupyter_console_1646669715337/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1645024702831/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1649765532372/work
jupyterlab-widgets @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_widgets_1647446862951/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1648737587623/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune @ file:///Users/runner/miniforge3/conda-bld/mistune_1635845001896/work
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1646999386773/work
nbconvert @ file:///Users/runner/miniforge3/conda-bld/nbconvert_1648465228596/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1649105480291/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1647377876077/work
numpy==1.22.3
opencv-python==4.5.5.64
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.0
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1649447152425/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1648873046251/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pycryptodome==3.14.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
pyOpenSSL==22.0.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
pyrsistent @ file:///Users/runner/miniforge3/conda-bld/pyrsistent_1649013444764/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1649055731282/work
rsa==4.8
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1638550740809/work
sqlparse==0.4.2
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1644872665635/work
terminado @ file:///Users/runner/miniforge3/conda-bld/terminado_1649104129126/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1645693042223/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1648827305171/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
urllib3==1.26.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
widgetsnbextension @ file:///Users/runner/miniforge3/conda-bld/widgetsnbextension_1647446968519/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
